requirements.txt
flask
google-generativeai
python-dotenv
flask-socketio
eventlet